Evaluate all files in validation workflow (fix)
This workflow was intended to run all files through pre-commit.
The input argument was not defined.
The input argument is now added as `additionalPreCommitArgs`
with a default value of `--all-files`.

https://pre-commit.com/\#pre-commit-run
'run all the hooks against all the files. This is a useful invocation if you are using pre-commit in CI.'